Analytics_Deployment/amls/model_deployment/requirements.txt
azure-storage-file-datalake
azureml-core